+++ src/main.rs
@@ -29,6 +29,7 @@ mod estimate;
 mod format;
 mod html;
 mod kde;
+mod message_formats;
 mod model;
 mod plot;
 mod report;
@@ -61,11 +62,6 @@ fn configure_log() {
     TermLogger::init(filter, Default::default(), TerminalMode::Stderr).unwrap();
 }
 
-// TODO: Add a section to the user guide
-// TODO: Add support for timelines & historical charts
-// TODO: Add machine-readable output
-// TODO: Notify burntsushi/critcmp about the internal file format change after I've added support for flat sampling
-// TODO: Look through github issues and implement them in cargo-criterion if sensible.
 // TODO: Write unit tests for serialization.
 
 /// Main entry point for cargo-criterion.
@@ -89,6 +85,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     let cli_report = configure_cli_output(self_config);
     let bencher_report = crate::report::BencherReport;
     let html_report = get_plotter(self_config)?.map(|plotter| crate::html::Html::new(plotter));
+    let machine_report = message_formats::create_machine_report(self_config);
 
     let mut reports: Vec<&dyn crate::report::Report> = Vec::new();
     match self_config.output_format {
@@ -99,6 +96,9 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     }
     if let Some(html_report) = &html_report {
         reports.push(html_report);
+    }
+    if let Some(machine_report) = &machine_report {
+        reports.push(machine_report);
     }
     let reports = crate::report::Reports::new(reports);
 
@@ -112,6 +112,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
                 &library_paths,
                 &reports,
                 &mut run_model,
+                self_config.message_format.is_some(),
             );
 
             if let Err(err) = err {
@@ -141,10 +142,10 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
 
 /// Configure and return a Report object that prints benchmark information to the command-line.
 fn configure_cli_output(self_config: &crate::config::SelfConfig) -> crate::report::CliReport {
-    let stdout_isatty = atty::is(atty::Stream::Stdout);
-    let mut enable_text_overwrite = stdout_isatty && !debug_enabled();
+    let stderr_isatty = atty::is(atty::Stream::Stderr);
+    let mut enable_text_overwrite = stderr_isatty && !debug_enabled();
     let enable_text_coloring = match self_config.text_color {
-        TextColor::Auto => stdout_isatty,
+        TextColor::Auto => stderr_isatty,
         TextColor::Never => {
             enable_text_overwrite = false;
             false